+++ tests/test_embedding.py
@@ -2,12 +2,11 @@
 
 import geomstats.backend as gs
 import geomstats.tests
+from geomstats.datasets.prepare_graph_data import HyperbolicEmbedding
 from geomstats.datasets.utils import load_karate_graph
-from geomstats.geometry.poincare_ball import PoincareBall
-from geomstats.learning.embedding_data import Embedding
 
 
-class TestEmbedding(geomstats.tests.TestCase):
+class TestHyperbolicEmbedding(geomstats.tests.TestCase):
     """Class for testing embedding."""
 
     def setUp(self):
@@ -20,16 +19,13 @@ class TestEmbedding(geomstats.tests.TestCase):
         context_size = 1
         karate_graph = load_karate_graph()
 
-        self.hyperbolic_manifold = PoincareBall(2)
-
-        self.embedding_class = Embedding(
-            data=karate_graph,
-            manifold=self.hyperbolic_manifold,
+        self.embedding_class = HyperbolicEmbedding(
+            graph=karate_graph,
             dim=dim,
             max_epochs=max_epochs,
             lr=lr,
-            n_negative=n_negative,
-            context_size=context_size)
+            n_context=context_size,
+            n_negative=n_negative)
 
     def test_log_sigmoid(self):
         """Test log_sigmoid."""
@@ -58,14 +54,14 @@ class TestEmbedding(geomstats.tests.TestCase):
             point, point_context, point_negative)
 
         expected_loss = 1.00322045
-        expected_grad = gs.array([[-0.16565083, -0.16565083]])
+        expected_grad = gs.array([-0.16565083, -0.16565083])
 
         self.assertAllClose(loss_value, expected_loss, rtol=1e-3)
-        self.assertAllClose(loss_grad, expected_grad, rtol=1e-3)
+        self.assertAllClose(gs.squeeze(loss_grad), expected_grad, rtol=1e-3)
 
-    @geomstats.tests.np_only
+    @geomstats.tests.np_and_pytorch_only
     def test_embed(self):
         """Test embedding function."""
         embeddings = self.embedding_class.embed()
-
-        self.assertTrue(self.hyperbolic_manifold.belongs(embeddings).all())
+        self.assertTrue(
+            self.embedding_class.manifold.belongs(embeddings).all())
